express connected to mongo and routes working
