+++ express_server/model/Trip.js
@@ -2,6 +2,9 @@ const mongoose = require('mongoose');
 const Schema = mongoose.Schema;
 
 let Trip = new Schema({
+  _id: {
+    type: Number
+  },
   state: {
     type: String
   },
