--- express_server/routes/book.routes.js
@@ -1,69 +1,0 @@
-const express = require('express');
-const app = express();
-
-const bookRoute = express.Router();
-let Trip = require('../model/Book');
-
-// Add Book
-bookRoute.route('/add-book').post((req, res, next) => {
-    Book.create(req.body, (error, data) => {
-    if (error) {
-      return next(error)
-    } else {
-      res.json(data)
-    }
-  })
-});
-
-// Get all Book
-bookRoute.route('/').get((req, res) => {
-    Trip.find((error, data) => {
-    if (error) {
-      return next(error)
-    } else {
-      res.json(data)
-    }
-  })
-})
-
-// Get Book
-bookRoute.route('/read-book/:id').get((req, res) => {
-    Book.findById(req.params.id, (error, data) => {
-    if (error) {
-      return next(error)
-    } else {
-      res.json(data)
-    }
-  })
-})
-
-
-// Update Book
-bookRoute.route('/update-book/:id').put((req, res, next) => {
-    Book.findByIdAndUpdate(req.params.id, {
-    $set: req.body
-  }, (error, data) => {
-    if (error) {
-      return next(error);
-      console.log(error)
-    } else {
-      res.json(data)
-      console.log('Book updated successfully!')
-    }
-  })
-})
-
-// Delete Book
-bookRoute.route('/delete-book/:id').delete((req, res, next) => {
-    Book.findByIdAndRemove(req.params.id, (error, data) => {
-    if (error) {
-      return next(error);
-    } else {
-      res.status(200).json({
-        msg: data
-      })
-    }
-  })
-})
-
-module.exports = bookRoute;
